+++ CHANGELOG.md
@@ -41,6 +41,15 @@
           builtin ``eval`` command can call these commands)
         - ``Module#_fullname`` method also supports theme
 
+#### Builtin
+
+- ``Error`` type objects maintain exit status
+    - now specify exit status to constructor
+    - add ``Error#status`` method for get exit status
+- add some options to builtin ``complete`` command
+    - ``-q``: does not completion candidates (but still set to ``COMPREPLY``)
+    - ``-s``: apppend space to completion candidate when number of candidates is 1
+
 #### Misc
 
 - add ``--html-lineno-table`` option to ``dscolorize``
@@ -116,9 +125,6 @@
       ```
       shctl set TRACE_ON_EXIT Null-Glob
       ```
-- ``Error`` type objects maintain exit status
-    - now specify exit status to constructor
-    - add ``Error#status`` method for get exit status
 
 #### API